各コマンドをファイルに分離
